// src/app/core/models/file.ts
export enum OvicMediaSourceTypes {
    local = 'local',
    serverFile = 'serverFile',
    vimeo = 'vimeo',
    youtube = 'youtube',
    googleDrive = 'googleDrive',
    encrypted = 'encrypted',
    serverAws = 'serverAws',
}
// ...
export interface OvicDocument {
    type: OvicDocumentTypes;
    source: OvicMediaSourceTypes;
    path: string | number;
    fileName?: string;
    preview?: boolean;
    _ext?: string | OvicDocumentTypes;
    download?: boolean;
    title?: string;
}

export enum OvicDocumentTypes {
    docx = 'docx',
    pptx = 'pptx',
    ppt = 'ppt',
    pdf = 'pdf',
    xlsx = 'xlsx',
    audio = 'audio',
    video = 'video',
    image = 'image',
    text = 'text',
    zip = 'zip',
}